# [core] split scheduler into smaller targets to improve build performance