--- a/BUILD.bazel
+++ b/BUILD.bazel
@@ -945,7 +945,11 @@ ray_cc_library(
 ray_cc_library(
     name = "scheduling_options",
     hdrs = ["src/ray/raylet/scheduling/policy/scheduling_options.h"],
-    deps = [":ray_common"],
+    deps = [
+        # TODO(Ziy1-Tan): ray_common is a giant dependency, replace it with sub-dependencies later
+        ":ray_common",
+        ":scheduling_context",
+    ],
 )
 
 ray_cc_library(
@@ -973,23 +977,28 @@ ray_cc_library(
     deps = [
         ":local_resource_manager",
         ":ray_common",
+        "//src/ray/protobuf:gcs_cc_proto",
+        "//src/ray/util:container_util",
+        "//src/ray/util:logging",
         "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_googletest//:gtest",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_googletest//:gtest_prod",
     ],
 )
 
 ray_cc_library(
     name = "cluster_resource_scheduler",
-    srcs = [
-        "src/ray/raylet/scheduling/cluster_resource_scheduler.cc",
-    ],
-    hdrs = [
-        "src/ray/raylet/scheduling/cluster_resource_scheduler.h",
-    ],
+    srcs = ["src/ray/raylet/scheduling/cluster_resource_scheduler.cc"],
+    hdrs = ["src/ray/raylet/scheduling/cluster_resource_scheduler.h"],
     deps = [
         ":cluster_resource_manager",
         ":composite_scheduling_policy",
         ":scheduler_internal",
+        "//src/ray/protobuf:gcs_cc_proto",
+        "//src/ray/util:logging",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_googletest//:gtest_prod",
     ],
 )
 
@@ -1007,18 +1016,33 @@ ray_cc_library(
         ":cluster_resource_scheduler",
         ":cluster_task_manager_interface",
         ":local_task_manager_interface",
+        ":scheduler_internal",
         ":scheduler_resource_reporter",
+        ":stats_lib",
+        "//src/ray/util:logging",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
     ],
 )
 
 ray_cc_library(
     name = "cluster_task_manager_interface",
     hdrs = ["src/ray/raylet/scheduling/cluster_task_manager_interface.h"],
+    deps = [
+        ":rpc_server_call",
+        "//src/ray/protobuf:node_manager_cc_proto",
+    ],
 )
 
 ray_cc_library(
     name = "local_task_manager_interface",
     hdrs = ["src/ray/raylet/scheduling/local_task_manager_interface.h"],
+    deps = [
+        ":ray_common",
+        ":scheduler_internal",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:flat_hash_set",
+    ],
 )
 
 ray_cc_library(
@@ -1028,9 +1052,12 @@ ray_cc_library(
     deps = [
         ":gcs_client_lib",
         ":ray_common",
+        "//src/ray/protobuf:gcs_cc_proto",
+        "//src/ray/protobuf:node_manager_cc_proto",
+        "//src/ray/util:logging",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
-        "@com_google_googletest//:gtest",
+        "@com_google_googletest//:gtest_prod",
     ],
 )
 
@@ -1038,7 +1065,10 @@ ray_cc_library(
     name = "affinity_with_bundle_scheduling_policy",
     srcs = ["src/ray/raylet/scheduling/policy/affinity_with_bundle_scheduling_policy.cc"],
     hdrs = ["src/ray/raylet/scheduling/policy/affinity_with_bundle_scheduling_policy.h"],
-    deps = [":scheduling_policy"],
+    deps = [
+        ":ray_common",
+        ":scheduling_policy",
+    ],
 )
 
 ray_cc_library(
@@ -1047,6 +1077,7 @@ ray_cc_library(
     hdrs = ["src/ray/raylet/scheduling/policy/bundle_scheduling_policy.h"],
     deps = [
         ":cluster_resource_manager",
+        ":ray_common",
         ":scheduling_context",
         ":scheduling_policy",
         ":scorer",
@@ -1061,10 +1092,13 @@ ray_cc_library(
         ":affinity_with_bundle_scheduling_policy",
         ":bundle_scheduling_policy",
         ":cluster_resource_manager",
+        ":hybrid_scheduling_policy",
         ":node_affinity_scheduling_policy",
         ":node_label_scheduling_policy",
         ":random_scheduling_policy",
         ":spread_scheduling_policy",
+        "//src/ray/util",
+        "//src/ray/util:container_util",
     ],
 )
 
@@ -1074,7 +1108,11 @@ ray_cc_library(
     hdrs = ["src/ray/raylet/scheduling/policy/hybrid_scheduling_policy.h"],
     deps = [
         ":scheduling_policy",
+        "//src/ray/util",
+        "//src/ray/util:container_util",
+        "@com_google_absl//absl/random",
         "@com_google_absl//absl/random:bit_gen_ref",
+        "@com_google_googletest//:gtest_prod",
     ],
 )
 
@@ -1092,14 +1130,21 @@ ray_cc_library(
     name = "node_label_scheduling_policy",
     srcs = ["src/ray/raylet/scheduling/policy/node_label_scheduling_policy.cc"],
     hdrs = ["src/ray/raylet/scheduling/policy/node_label_scheduling_policy.h"],
-    deps = [":scheduling_policy"],
+    deps = [
+        ":scheduling_context",
+        ":scheduling_policy",
+    ],
 )
 
 ray_cc_library(
     name = "random_scheduling_policy",
     srcs = ["src/ray/raylet/scheduling/policy/random_scheduling_policy.cc"],
     hdrs = ["src/ray/raylet/scheduling/policy/random_scheduling_policy.h"],
-    deps = [":scheduling_policy"],
+    deps = [
+        ":scheduling_policy",
+        "//src/ray/util",
+        "//src/ray/util:container_util",
+    ],
 )
 
 ray_cc_library(
@@ -1109,6 +1154,8 @@ ray_cc_library(
     deps = [
         ":hybrid_scheduling_policy",
         ":scheduling_policy",
+        "//src/ray/util",
+        "//src/ray/util:container_util",
     ],
 )
 
@@ -1125,7 +1172,9 @@ ray_cc_library(
     hdrs = ["src/ray/raylet/scheduling/scheduler_resource_reporter.h"],
     deps = [
         ":local_task_manager_interface",
+        ":ray_common",
         ":scheduler_internal",
+        "@com_google_absl//absl/container:flat_hash_map",
     ],
 )
 
@@ -1133,7 +1182,7 @@ ray_cc_library(
     name = "scheduling_policy",
     hdrs = ["src/ray/raylet/scheduling/policy/scheduling_policy.h"],
     deps = [
-        ":scheduling_context",
+        ":ray_common",
         ":scheduling_options",
     ],
 )