--- a/src/ray/raylet/scheduling/policy/spread_scheduling_policy.cc
+++ b/src/ray/raylet/scheduling/policy/spread_scheduling_policy.cc
@@ -17,7 +17,6 @@
 #include <functional>
 
 #include "ray/util/container_util.h"
-#include "ray/util/util.h"
 
 namespace ray {