--- a/src/ray/raylet/scheduling/scheduling_policy.cc
+++ b/src/ray/raylet/scheduling/scheduling_policy.cc
@@ -19,8 +19,6 @@
 #include <utility>
 #include <vector>
 
-#include "ray/util/container_util.h"
-
 namespace ray {
 
 namespace raylet_scheduling_policy {