<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

<!-- Please give a short summary of the change and the problem this solves. -->

## Related issue number
Closes #51634 

Before:
```
INFO: Elapsed time: 600.027s, Critical Path: 94.84s
INFO: 3687 processes: 460 disk cache hit, 1400 internal, 1827 linux-sandbox.
INFO: Build completed successfully, 3687 total actions
```
After:
```
INFO: Elapsed time: 440.950s, Critical Path: 89.94s
INFO: 3932 processes: 1647 internal, 2285 linux-sandbox.
INFO: Build completed successfully, 3932 total actions
```

Reproduce steps:
```
bazel clean
sudo sync; sudo sh -c "echo 3 > /proc/sys/vm/drop_caches"
bazel build //:scheduler
```

About **26.5%** compile time reduction.

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
    - [ ] I've added any new APIs to the API Reference. For example, if I added a
           method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(