@dentiny 2025-03-24
> About 21.8% compile time reduction.

Thanks for the benchmark! Curious what target are you building?