@Ziy1-Tan 2025-03-25
> > About 21.8% compile time reduction.
> 
> Thanks for the benchmark! Curious what target are you building?

The target //:scheduler only. I have updated some reproduce steps.